Fix: kill all openocd children even on interrupts

+++ scripts/debug.sh
@@ -31,19 +31,41 @@ list_descendants() {
   echo "$children"
 }
 
+
+# Catch interrupts
+exitfn () {
+    trap SIGINT
+    if [ -n "${ocdpid}" ]; then
+      echo "Killing pid $ocdpid and all children"
+      kill $(list_descendants $ocdpid)
+      # killall openocd
+    fi
+    exit 
+}
+trap "exitfn" INT
+
+
+# Create a temp file for OpenOCD output
 TMPFILE=/tmp/stm32-ocd.out
 if [ -f ${TMPFILE} ]; then
   truncate -s0 ${TMPFILE}
 else
   touch ${TMPFILE}
 fi
+
+
+# Run OpenOCD in background
 make openocd OPENOCD_ARGS="--log_output ${TMPFILE}" &
 ocdpid=$!
 
+
+# Wait for flash the chip
 tail -f ${TMPFILE} | grep -q '\*\* Resetting Target \*\*'
 
+
+# Run debugger
 make gdb
-exitstatus=$?
-echo "Killing pid $ocdpid and all children"
-kill $(list_descendants $ocdpid)
-exit $exitstatus
+
+
+# Restore trap and exit!
+exitfn
